WIP: scraper and logger

+++ internal/db/model_metadata.go
@@ -1,6 +1,6 @@
 package db
 
 type Timestamp struct {
-	CreatedAt int `json:"created_at"`
-	UpdatedAt int `json:"updated_at"`
+	CreatedAt int64 `json:"created_at"`
+	UpdatedAt int64 `json:"updated_at"`
 }
